backend/tadpool/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
boto3==1.38.8
botocore==1.38.8
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==44.0.3
des==1.0.6
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-rest-knox==5.0.2
django-storages==1.14.6
djangorestframework==3.16.0
et_xmlfile==2.0.0
filelock==3.18.0
firebase-admin==6.8.0
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.6
jmespath==1.0.1
msgpack==1.1.0
multidict==6.2.0
numpy==2.2.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pipenv==2025.0.1
platformdirs==4.3.7
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pygame==2.6.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.31.0
rsa==4.9.1
s3transfer==0.12.0
setuptools==80.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.1
utils==1.0.2
virtualenv==20.30.0
yarl==1.18.3
